autopartscentral/requirements.txt
Django==1.9.5
django-user-accounts==1.3.1
pinax-webanalytics==2.0.1
pinax-eventlog==1.1.1
pinax-theme-bootstrap==7.8.0
django-widget-tweaks==1.4.1